add the `supportsZeroMem` type trait (#1307)
## Summary

The type is intended for libraries to figure out whether `zeroMem`
leaves a location in its "default-initialized" state, something that's
useful to know for optimization or correctness purposes in generic
code.

## Details

Without the type trait, either approximation via `is`, or a type-
inspection macro would have to be used. A built-in type trait is
better than a macro, since it's both more efficient and usable in
the `system` module.

The only types not supporting `zeroMem` for initialization are:
* types that don't have a zero-default (such as `range` types, enums
  not starting at zero, `.requiresInit` types, not-nil pointers, etc.)
* object types that have a type header field (non-pure object types
  that can be inherited from)
* incomplete package-level object types
* aggregate types containing the former three

The `nimskullHasSupportsZeroMem` conditional symbol is added so that
support for type-trait can be detected at compile-time.