always clear slots when growing `seq` with `setLen` (#1306)
## Summary

Growing a `seq[T]` with `setLen` after shrinking it now doesn't retain
the slots' previous state, default-initializing them instead. This was
already the behaviour for destructor-having element types, but not for
other types.

## Details

For efficiency, two optimizations are used:
* the slots memory is only default-initialized when *growing* with
  `setLen`, not when shrinking
* if the element type supports it, a single `zeroMem` call is used for
  clearing the new slots after growing

For types that don't support `zeroMem`, `default(T)` is assigned to
every new slot, which fixes type header for non-pure inheritable
objects not being initialized (and `of` thus crashing).

The `seq` implementation procedures have to assume that after resizing,
the memory of the new slots is in an undefined state, which means that
no hook calls (`=copy` or `=sink`) must be used. This is guaranteed by
performing the assignments within `.nodestroy` procedures.

Since slot assignments never expand to hook calls now, the payload
memory doesn't have to be zero'ed on allocation anymore.