internal: use AST for dynlib loaders
Summary
=======

Instead of directly producing MIR code for the dynlib loaders, first
produce an AST that is then translated to the MIR, reducing the amount
of code that needs to be adjusted when the MIR changes.

Details
=======

Generating an intermediate AST means that `mirgen` can fully be
leveraged for the translation, instead of only partially, which ensures
proper translation even when the MIR changes.

The `mirgen-generateCode` overload for appending to an existing
`MirBuilder` is now obsolete and thus removed. In addition, the pre-
and post-MIR-translation body for dynlib loaders can now be echoed via
the `-d:nimShowMirInput` and `-d:nimShowMir` debug facilities.