fix(js): slow code generation
Summary
=======

Replace an algorithm - used by the JavaScript code generator -
potentially taking an exponential amount of time with one taking linear
time. In some edge cases, the previous algorithm was so slow that the
compiler would effectively get stuck.

Details
=======

`jsflow` is responsible for translating the goto-based control-flow of
the CGIR into JavaScript control-flow (`break`, `try`, etc.). This
requires reordering the initial list of block starts (openings) and
ends such that all openings are paired with their corresponding end.

Previous Algorithm
------------------

The algorithm used for reordering the list works by:
1. start with the last slot in the list
2. shift the block/try opening in the slot backwards (towards the
   start) until it is on the same nesting depth as its corresponding
   end
3. repeat 2 until a fixpoint is reached (the item in the slot no longer
   requires shifting)
4. repeat the previous two steps with the previous slot

Consider the following (where `n` is a value >= 2):
```
block { // 0
  block { // n
    ... // (n - 2) .. 1
    block { // n - 1
    } // 0
  } // 1
  ...
} // n
```

In the situation show above, reaching a fixpoint for the slot
originally containing the `n - 1` block requires `3 * (2 ^ (n - 2))`
iterations/shifts (exponential time).

For large `n`, the time spent on sorting grew so large that the
compiler would appear to be effectively stuck.

New Algorithm
-------------

The new algorithm requires `n` number of shifts in the worst-case,
where `n` is the number of labels in the input CGIR.

It works by first computing an order value for every label, based on
the position of the label. The relevant openings (block and try) are
then reordered such that they're only enclosed by openings with a
higher order value.