jsgen: defer `arrayConstr` registration
If `createVar` is called from `defineGlobal`, no new entities must be
registered with the `MirEnv`. To make the `arrayConstr` procedure is
still registered, if `createVar` is called from `defineGlobal`, the
`useMagic` call is effectively deferred until the `genProc` or
`genPartial` call.

It's currently guaranteed that discovery of a global (which invokes
`defineGlobal`) is followed by code generation of at least one
procedure.