fix(js): compiler crash with globals (#1314)
## Summary

Fix a compiler crash that happened when having globals with type
`array[N, T]`, where `N` is greater than 32 and `T` is a non-primitive
type. Only the JavaScript backend was affected.

## Details

Defer adding the `arrayConstr` compilerproc to the MIR environment if
not within a proper procedure context. This prevents modifications of
the MIR environment from within `defineGlobal` (which calls
`createVar`), which is what caused the assertion in `backends.flush` to
fail.

Registration is deferred until the procedure is code generated, which
works due to discovery of a global currently being guaranteed to
precede code generation of some procedure.