mirgen: account for unreachable expressions
A statement list expression can have noreturn statements within
(such as `raise`), in which case the expression is unreachable. The
expression may depend on locals defined within the unreachable parts
of the statement list, so translating it could crash the compiler (due
to a missing entry in the `localsMap` table).

Where easy to do, non-returning expressions are handled by omitting
the statement they're part of. In all other cases, the trailing
expression is translated into a dummy temporary.