injectdestructors: eliminate unnecessary destroys
If the target location is guaranteed to not contain a value when
computing the destroy operation, the destroy operation can be elided.
This can happen if the location is either never assigned to or moved
out of on all paths.

Since removing instructions from the DFG is not possible, the
`opNone` opcode is introduced. All eliminated destroy operations are
then replaced with `opNone` in the DFG.