fix csource compilation with many modules
Summary
=======

Make building the csource compiler on Windows via the `build.bat` batch
script work when the link command length exceeds the maximum command
line length due to a large number of object files.

Details
=======

On Windows, everything beyond 8192 characters is cut off from the
command-line arguments, leading to not all arguments reaching the
linker.

Instead of directly appending the object file paths to a the link
command, `build.bat` now writes the paths to the temporary
"linker_args.rsp" file, which is then passed to gcc as a "response
file".

For simplicity, this is done even when the command-line would not
exceed the maximum length.