mir: implement the new control-flow primitives (#1281)
## Summary

Replace the legacy control-flow primitives with the new ones outlined
in the documentation. This is a refactor for the most part, but - as a
side effect - a bug is fixed too, and destructors are no longer called
on not-yet-alive locations.

Fixes #1303.

## Details

The new control-flow primitives are those documented in the `mir.rst`
file -- the node kinds for the legacy ones are all removed. They're the
same as those used by the CGIR, removing the largest remaining
difference between the two IRs, preparing for the CGIR's removal.

In broad:
* translating the `try`/`block`/`break` control-flow to the goto-based
  control-flow now happens during AST -> MIR translation (`mirgen`)
  instead of during MIR -> CGIR translation (`cgirgen`)
* elimination of unreachable code now happens during AST -> MIR
  translation already
* all necessary destroy operations are injected during AST -> MIR
  translation
* the "destructor injection" pass now only collapses sink operations
  and eliminates destroy operations that turn out to be redundant

Adding new finally sections after the initial MIR production would
require patching all intercepted gotos and checked calls, which is
complex and hurts composition of the pass, hence this approach being
decided against.

### Multiple `finally`s Per Scope

Instead of using a single `finally` per scope (or none at all),
multiple ones are used, depending on the placement of unstructured
scope exits. As a consequence:
```nim
block:
  var x = Object()
  if cond:
    break # no longer invokes the destructor for `y`
  var y = Object() # destruction doesn't require a finally
```

This generally enables better code generation, since locals destroyed
in `finally` sections don't have to conservatively be zero'ed on scope
entry anymore.

### Other Changes

* the MIR pretty-printer (`mir/utils.nim`) is updated to render the new
  node kinds
* `transf` no longer wraps `while` loop bodies in a `block`, which was
  previously required for proper destruction of locals defined as part
  of the body
* `mirtrees.numArgs` now only returns the number of arguments in a call
  tree, not the number of child nodes
* repeated are `pirMat`/`pirMatCursor` nodes are guarded against during
  PMIR production
* the `mnkDefUnpack` node kind is obsolete now that destroy operations
  are placed by `mirgen` -- it's removed

### Test Changes

* tests using now-removed MIR node kinds are changed to use different
  ones
* some assertions in now-working tests are updated
* the `--expandArc`-using tests are updated to the changes in MIR
  rendering