fix `result` not being cleaned up when procedure raises (#1326)
## Summary

The `result` variable is now destroyed when the procedure exits due to
a raised exception, fixing a memory/resource leak. All backends were
affected.

Fixes https://github.com/nim-works/nimskull/issues/1305.

## Details

* the `result` is destroyed within a special try-finally that only
  intercepts exceptional control-flow
* the new `errorOnly` field designates a `bkTryFinally` as only
  intercepting exceptional control-flow
* `mirgen` handles emitting the destroy operation and injecting the
  `try`

When the `result` variable is not initialized on any control-flow paths
leading to the destroy operation, the destroy is elided by the "destroy
elimination" pass; a test (`tno_destroy_for_empty_result.nim`) is added
to ensure this works.