lower integer overflow checks with MIR pass (#1325)
## Summary

Lower the integer overflow checks with an MIR pass, instead of as part
of C code generation. This is mainly a refactoring, with the goal of
shrinking down the C code generator. As a side-effect, overflow checks
are very slightly faster.

## Details

The MIR produced by the lowering is similar to what the C code
generator previously output, but with one major difference: integer
operands smaller than the target's integer size now use dedicated
compilerprocs instead of first converting to `int`, performing the
checked arithmetic, and then comparing against the integer bounds.

The main reason for this change is that it makes the lowering pass
simpler and requires less MIR code, but it also has the benefit of
producing more efficient code. Run-time behaviour is not affected.

No dedicated gcc built-ins for overflow-checked arithmetic with
`short int` (int16) and `signed char` (int8) exists, so all checked
arithmetic compilerprocs now map to the generic gcc built-ins.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>