properly check index operand against array bounds
If the low/high index of an array cannot be represented by the index
operand's type (`T`), comparing the boundary indices against the index
operand the can lead to both false positive and false negative index
check results, due to C's integer promotion applying because of the
type mismatch.

For example:
```nim
  # false negative:
  let i = low(int)
  var arr: array[uint(high(int))..succ(uint(high(int))), int]
  discard arr[i]
  # false positive:
  let i = 1
  var arr: array[-1..1, int]
  discard arr[i]
```

So that the comparison operands' types match, the respective boundary
index is first converted (at compile-time) to `T`. If the boundary
index value is not representable with `T`, the comparison can safely
be skipped. For example, if `T` is `uint` and the array's lower index
boundary is -1, testing whether the index operand is >= -1 is
unnecessary, because a `uint` is guaranteed to be >= 0.