remove unnecessary `ccgutils` dependency (#1330)
## Summary

Remove an obsolete usage of `ccgIntroducedPtr` from sem, breaking an
import dependency between sem and the code generator, and allowing for
`ccgIntroducedPtr` to move back to `ccgtypes`.

## Details

It was originally possible for the arguments to a copy hook overlapping
in memory (a parameter aliasing violation), resulting in incorrect
behaviour when copying variant objects if they do.

ed126794b6af8b39960a7acc603cee1abc73da4a worked around this by checking
whether both parameters were (hidden) pointers to the same location,
skipping the copy if they are, with `ccgIntroducedPtr` used to check
whether the source parameter is passed as via a pointer indirection.

This is a layering violation, and also no longer necessary, since
`injecthooks` ensures that the immutable argument doesn't refer to the
same memory location as the mutable one -- the workaround in
`liftdestructors` can thus be removed.