mir: remove statement nesting (#1331)
## Summary

Scopes are now no longer subtrees that can have other statements nested
within them. This removes the last of case of statement nesting from
the MIR, meaning that all bodies are now made up of flat statement
lists.

## Details

Statement nesting complicates tree traversal, and - more importantly -
prevents the eventual removal of the `mnkEnd` nodes. Without `mnkEnd`
nodes, the number of sub-nodes is going to be explicitly stored in the
subtree nodes, which cannot efficiently work in the `Changeset`
architecture when statements can be nested.

Construction of MIR trees is updated accordingly (`subTree` cannot be
used for wrapping statements in a scope; the `scope` template is used
instead), as well as tree traversal.