transf: remove unreachable code (#1329)
## Summary

Eliminate unreachable statements and expression during `transf`,
instead of in `mirgen`. This:
* removes some workarounds from `mirgen`
* improves the quality of unreachable code elimination
* fixes a compiler crash related to non-returning expressions

## Details

Statement lists can have trailing unreachable code, and expressions can
therefore be non-returning (e.g., `(;break; 1)`).

`mirgen` made an attempt at eliminating both, but the approach used
was indirect and didn't handle expressions well, leading to crashes for
code such as `f((;return; var x = 0; 1), x)`, because the `var x`
definition wasn't translated, but the subsequent usage of `x` was.

Removing unreachable statements/expressions is now done by an AST-based
pass that's run after early `transf` (so that iterator inlining took
place and the AST is less complex) but before the `closureiters`
transformation. The pass is implemented by the `unreachable_elim`
module.

Beyond fixing the crashes, this has the additional benefits that:
* the `closureiters` transformation doesn't take unreachable code into
  account
* `mirgen` can detect `finally` clauses that have no structured exit
  and emit better code for them. The handling for the no-structured-
  exit case already existed (`doesntExit`), just the detection was
  missing
* more straightforward MIR translation, using less conditionals

A test for the various possible scenarios of unreachable expressions/
statements is added.