cbackend: implement the basic write-to-disk logic
It works much like the previous version, but with more generalized
support for header files. Compare to before, all the write-to-disk
management is now fully handled by the orchestrator, not the code
generator (i.e., `cgen`).

The compiler compiles again (but the result cannot compile the
compiler, for obvious reasons).