mir: add the `mnkBinding` node
It has the same role as the `cnkBinding` node in the CGIR, that is, to
group a field name and value expression in an object construction
expression together.

Previously, the `mnkField` and expression nodes loosely followed each
other, making `mnkObjConstr` subtrees heterogenous, which complicates
some traversal (or at least it will, once more traversal is tree-
length-based).