mirconstr: adjust the builder
A `MirBuffer` instance keeps the track of the number of the subnodes in
the currently constructed tree, updating the `len` field in the root
node on subtree completion.

The value is saved on the stack when starting a nested subtree. For
manual subtree construction (i.e., not using the `subTree` template),
the `start` and `finish` procedures are provided.

Since the builder manages updating the length field, manually providing
the number of nodes is unnecessary (`mirgen` is updated accordingly).