macros: fix compiler crash with `parseExpr` (#1339)
## Summary

Fix a bug with `parseExpr` error handling that caused a compiler crash
when:
* an empty string was passed as the argument
* parsing the string returned more than one statement block

## Details

A `SemReport` was used instead of the correct `VMReport`. Since
`rvmOpcParseExpectedExpression` is not a valid report kind for
`SemReport`, the assertion in `wrap` failed and the compiler crashed.

The `ttryparseexpr` test is cleaned up and extended with two test cases
for the cases that trigger the "expected expression" case, ensuring
that previously faulty branch now has test coverage.