vm: don't use ref-counted cell for globals
Summary
=======

Store a handle to a normal memory cell in the `globals` table, instead
of a handle to a ref-counted cell.

Details
=======

VM globals are normal locations that exists as long as the VM
environment does; they don't need to be reference-counted, which they
also never were.

Storing a handle to the memory cell in the global table directly
removes and unnecessary indirection through `HeapSlotHandle`, and thus
speeds up loading a global a tiny bit.