fix: wrong code generation for synthesized object hooks (#1338)
## Summary

Fix a bug with synthesis of lifetime-tracking hooks for object types
using inheritance, which showed up as a C compiler error when using
recent versions of clang or gcc.

Fixes https://github.com/nim-works/nimskull/issues/1337.

## Details

When synthesizing the hooks for object types using inheritance, the
destination accessor expression wasn't wrapped in an object conversion
prior to emitting the parent object handling.

Without the conversion, the synthesized hook passed along its own
parameter to the call of the super type's hook as-is. This is wrong
in general (the argument's type doesn't match the parameter's), but
didn't cause concrete issues until gcc and clang started to disallow
implicit conversions between incompatible pointer types.