fix: crash when defining call/dot operators (#1341)
## Summary

Fix the compiler crashing when defining dot or call operator overloads
and the respective experimental feature is not enabled.

## Details

The formatting of the "call/dot operator not enabled" error messages
expects a symbol, but none was provided, resulting in an NPE.

For both "call/dot operator not enabled" diagnostics, when turning
them into reports, the symbol is now taken from the wrong node and
passed into the report.