cbackend: sketch out the data layout/ownership
Everything only needed within a single module is stored in `BModule`,
things that are shared are stored globally (in `BModuleList`).

This keeps the scopes of local entities small, and will make it easy to
free memory early (by destroying a `BModule` instance once the C code
for it has been generated).