macros: keep instantiation parameters with `getType` (#1346)
## Summary

`getType` now keeps instantiation parameters, which were previously
irrecoverably lost. This also allows retrieving the instantiation
parameters when only having access to an alias of the instantiated
type.

Fixes #1223.

## Details

Expanding `tyGenericInst` to `(BracketExpr <Callee> ...)` is the only
way of keeping the type information intact, since `tyGenericInst` do
not have symbols attached to them (in which case translating to an
`nnkSym` would be preferable, as it'd keep the AST flat).

Existing tests are adjusted.

### Aliases

A second goal of changing `getType` behaviour for `tyGenericInst` is to
allow retrieving the instantiation parameters in the presence of
type aliases. Given:
```nim
type
  Generic[A, B] = object
  Alias = Generic[int, string]
```

it was previously not possible to retrieve the instantiation
parameters for `Generic[int, string]` when given the expression
`Alias`, but now it is (with `getType`).

### Other

The internal `enumutils.enumFullRange` macro relied on the old
`getType` behaviour. To restore its functionality, it now uses
`getTypeImpl`.