sem: don't analyze dependent operands to `static` (#1345)
## Summary

Makes instantiations such as `Generic[someProc(T)]` work, where
`T` is some generic parameter (and receiving type parameter is a
`static` one) -- the `someProc(T)` expression is typed once `T` is
substituted, making the behaviour consistent with `array` and `range`.

## Details

Fully analyzing (i.e., with `semExpr`) arbitrary expressions that might
reference unresolved type variables generally doesn't work. Operands to
generic invocations were eagerly analyzed (in `matchesAux`), usually
resulting in errors when the operand needs to be a `static` value.

Disallowing expressions dependent on unresolved type variables in this
context would be a regression, since some expressions are special-cased
to work (such as `sizeof(T)`).

To address the problem, before typing the operand, it's first checked
whether it depends on unresolved type variables. If it does, the
expression is treated as a generic expression, with a `tyFromExpr`
assigned as its type -- otherwise it's analyzed as usual.

If the operand has an unknown type (`tyFromExpr`), it's wrapped in a
conversion to a type derived from the generic parameter's constraint,
to make sure the operand types later. If no concrete type can be
derived from the constraint, a type mismatch is reported.

To not convolute `matchesAux` further, the argument matching where the
callee is a `tyGenericBody` is moved into the new `matchesType`. It
does largely the same as `matchesAux`, but with everything not
applicable to types removed (such as the varargs handling).

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>