nversion: read version from text file
Instead of statically encoding the version as a structure, read it from
a text file. This allows tools to modify the version string without
creating invalid NimSkull, and allow for easier rearrangement of the
code in the future.

Additional validation has also been added to prevent the version string
from containing invalid characters per SemVer specification.

The version number is now stored in `compiler/version.txt`.