fix incorrect typing in closure-iterator C code-gen
Summary
=======

Fix the C code generator emitting incorrectly typed code for closure
closure-iterators, which recent stricter C compilers report errors for.

Details
=======

When computing the environment for closure iterators, and the closure
iterator closes over some outer locals, there's already a symbol in the
hidden parameter slot, using the environment type of the enclosing
routine.

The symbol is retyped during closure iterator environment computation,
but the *node* referencing it was not updated. `cgen` used the MIR
locals type for the definition, but the *node's* type (which was still
the enclosing procedure's environment type) for the cast, resulting in
a pointer type mismatch.

The node referencing the symbol is now also updated, and an assertion
to make sure the types match is added in `cgen`.