mirbodies: track next label in `MirBody` (#1365)
## Summary

Track the ID to use for new labels in `MirBody`. This greatly reduces
the overhead of `Changeset` creation, since the source body doesn't
have to be scanned.

## Details

* track the next label ID with `MirBody.nextLabel`
* on creation, the `Changeset` next label ID is initialized from the ID
  stored with the `MirBody`
* add the `createBody` routine for creating a proper `MirBody` from a
  `MirBuilder` instance
* use `createBody` to replace manual `MirBody` construction

### Performance

Compile times are effectively unaffected. A small improvement of a few
milliseconds is measurable, but it's insignificant. However, using more
batches in the future will make the improvement more pronounced.