ci: use concurrency to cancel obsolete runs
We used to do this with skip-duplicate-actions, but this has not been
working since we tighten up the permissions given to workflows.

Replace this with the native `concurrency` feature, which allows us to
cancel old runs without needing a 3rd-party actions. This also works for
PRs as well, so if someone push a new commit to their PR, previous CI
run for that PR will be cancelled.