add changelog builder for release publishing (#1377)
## Summary
Add changelog builder to the publishing workflow. This allows us to
provide richer release information to users.

## Details
The changelog builder will tally all PRs merged between the previous tag
and the latest tag, then generate a categorized changelog for them. The
configuration for the builder can be found at  `.github/changelog.json`
.

In addition to this,  `/generate-changelog`  PR command has been added,
which allows maintainers to generate a changelog based on the new
configuration within a given PR.

Usage:

/generate-changelog [from=<tag>] [to=<tag>] [no_prerelease=<true|false>]

Where:

-  `from` : Specify the previous tag. If not provided, will be the first
tag older than  `to` .
-  `to` : Specify the latest tag. If not provided, will be the latest
tag in this repository.
-  `no_prerelease` : Only applies when  `from`  is not provided. If set
to  `true` ,  `from`  is selected such that the older tag is /not/ a
pre-release. Defaults to  `false` .