ccgtypes: fix compiler crash
The `m.fields[id]` table lookup borrows, even though in this specific it
cannot (since `m` is a `ref` and its content possibly modified
afterward by `genTypeInfoV1`). If `genTypeInfoV1` modified the fields
table, the formatting logic would crash. To work around this compiler
bug, a lookup procedure that copies the string is used.