sem, pragmas: fix fatal pragma (#1386)
## Summary

Using  `{.fatal:"...".}`  did not show any message to the user and it
didn't stop compilation.
This commit fixes both issues.

## Details

-  `adSemFatalError`  now carries a message just like 
`adSemCustomUserError` .
- All reports of `fatal` severity abort compilation.
- Testament now also consider fatal errors to be errors.
- A simple test case has been included.
-  `isCompilerFatal`  has been removed as it seems it lost its purpose.
The check for  `rextCmdRequiresFile`  seems to be unnecessary. A fatal
report is produced elsewhere (the only place where 
`rextCmdRequiresFile`  is used) before this function is ever called, so
it was redundant and it doesn't break existing code.